rework based on review comments
